Move HandsContext to top level
